+++ src/utilities/attack.py
@@ -290,81 +290,3 @@ class Attack:
         )
 
         return bishop_attacks | rook_attacks
-
-    @staticmethod
-    def check_squares_attacked(app, square, side):
-        # Leaper pieces
-        # Attacked by white pawns
-        if (side == SIDES["white"]) and (
-            app.table_manager.pawn_attack_table[SIDES["black"]][square]
-            & app.bitboard_manager.bitboards[PIECES["P"]]
-        ):
-            return True
-
-        # Attacked by black pawns
-        if (side == SIDES["black"]) and (
-            app.table_manager.pawn_attack_table[SIDES["white"]][square]
-            & app.bitboard_manager.bitboards[PIECES["p"]]
-        ):
-            return True
-
-        # Attacked by knight
-        if (
-            app.bitboard_manager.bitboards[PIECES["N"]]
-            if side == SIDES["white"]
-            else app.bitboard_manager.bitboards[PIECES["n"]]
-        ) & (app.table_manager.knight_attack_table[square]):
-            return True
-
-        # Attacked by king
-        if (
-            app.bitboard_manager.bitboards[PIECES["K"]]
-            if side == SIDES["white"]
-            else app.bitboard_manager.bitboards[PIECES["k"]]
-        ) & (app.table_manager.king_attack_table[square]):
-            return True
-
-        # Slider pieces
-        # Attacked by bishop
-        occupancy = app.bitboard_manager.occupancies[SIDES["all"]]
-        attack_table = app.table_manager.bishop_attack_table
-        attack_masks = app.table_manager.bishop_attack_masks
-
-        if (
-            app.bitboard_manager.bitboards[PIECES["B"]]
-            if side == SIDES["white"]
-            else app.bitboard_manager.bitboards[PIECES["b"]]
-        ) & Attack.get_bishop_attack_masks(square, occupancy, attack_table, attack_masks):
-            return True
-
-        # Attacked by the rook
-        occupancy = app.bitboard_manager.occupancies[SIDES["all"]]
-        attack_table = app.table_manager.rook_attack_table
-        attack_masks = app.table_manager.rook_attack_masks
-
-        if (
-            app.bitboard_manager.bitboards[PIECES["R"]]
-            if side == SIDES["white"]
-            else app.bitboard_manager.bitboards[PIECES["r"]]
-        ) & Attack.get_rook_attack_masks(square, occupancy, attack_table, attack_masks):
-            return True
-
-        # Attacked by the queen
-        occupancy = app.bitboard_manager.occupancies[SIDES["all"]]
-        attack_tables = (
-            app.table_manager.bishop_attack_table,
-            app.table_manager.rook_attack_table,
-        )
-        attack_masks = (
-            app.table_manager.bishop_attack_masks,
-            app.table_manager.rook_attack_masks,
-        )
-
-        if (
-            app.bitboard_manager.bitboards[PIECES["Q"]]
-            if side == SIDES["white"]
-            else app.bitboard_manager.bitboards[PIECES["q"]]
-        ) & Attack.get_queen_attack_masks(square, occupancy, attack_tables, attack_masks):
-            return True
-
-        return False
